requirements.txt
pyautogen>=0.2.30
openai>=1.30.0
python-dotenv>=1.0.0
rich>=13.7.0
pydantic>=2.0.0
